removed stage order from opening
